+++ src/main/kotlin/no/nav/bidrag/behandling/dto/v1/behandling/BehandlingDto.kt
@@ -22,7 +22,7 @@ data class OppdaterManuellVedtakResponse(
 
 data class OppdaterManuellVedtakRequest(
     val barnId: Long,
-    val vedtaksid: Long?,
+    val vedtaksid: Int?,
 )
 
 data class ManuellVedtakResponse(
@@ -30,7 +30,7 @@ data class ManuellVedtakResponse(
 )
 
 data class ManuellVedtakDto(
-    val vedtaksid: Long,
+    val vedtaksid: Int,
     val barnId: Long,
     val fattetTidspunkt: LocalDateTime,
     val virkningsDato: LocalDate,
@@ -72,7 +72,7 @@ data class VirkningstidspunktDtoV2(
     @Schema(description = "Løpende opphørsvedtak detaljer. Er satt hvis det finnes en vedtak hvor bidraget er opphørt")
     val eksisterendeOpphør: EksisterendeOpphørsvedtakDto? = null,
     @Schema(description = "Manuell vedtak valgt for beregning av aldersjustering")
-    val grunnlagFraVedtak: Long? = null,
+    val grunnlagFraVedtak: Int? = null,
     val kanSkriveVurderingAvSkolegang: Boolean = false,
     val manuelleVedtak: List<ManuellVedtakDto> = emptyList(),
 ) {
